Cleanup parent's playback mess
