+++ lib/membrane/core/parent/lifecycle_controller.ex
@@ -3,7 +3,7 @@ defmodule Membrane.Core.Parent.LifecycleController do
   use Bunch
   use Membrane.Core.PlaybackHandler
 
-  alias Membrane.Core
+  alias Membrane.{Core, Sync}
 
   alias Core.{
     Parent,
@@ -20,12 +20,39 @@ defmodule Membrane.Core.Parent.LifecycleController do
   require PadModel
   require Membrane.PlaybackState
 
-  # TODO delete this type as it's already defined
-  @type handlers :: %{
-          action_handler: module(),
-          playback_controller: module(),
-          spec_controller: module()
-        }
+  @impl PlaybackHandler
+  def handle_playback_state(old, new, state) do
+    children_data =
+      state
+      |> Core.Parent.ChildrenModel.get_children()
+      |> Map.values()
+
+    children_pids = children_data |> Enum.map(& &1.pid)
+
+    children_pids
+    |> Enum.each(&Parent.change_playback_state(&1, new))
+
+    :ok = toggle_syncs_active(old, new, children_data)
+
+    state = %{state | pending_pids: children_pids |> MapSet.new()}
+
+    if children_pids |> Enum.empty?() do
+      {:ok, state}
+    else
+      PlaybackHandler.suspend_playback_change(state)
+    end
+  end
+
+  @impl PlaybackHandler
+  def handle_playback_state_changed(old, new, state) do
+    callback = PlaybackHandler.state_change_callback(old, new)
+
+    if new == :stopped and state.terminating? do
+      Message.self(:stop_and_terminate)
+    end
+
+    CallbackHandler.exec_and_handle_callback(callback, state.handlers.action_handler, [], state)
+  end
 
   def handle_spec(spec, state) do
     with {{:ok, _children}, state} <-
@@ -90,16 +117,6 @@ defmodule Membrane.Core.Parent.LifecycleController do
     )
   end
 
-  @impl PlaybackHandler
-  def handle_playback_state(old, new, state) do
-    Parent.handle_playback_state(old, new, state)
-  end
-
-  @impl PlaybackHandler
-  def handle_playback_state_changed(old, new, state) do
-    Parent.handle_playback_state_changed(old, new, state)
-  end
-
   # TODO get pending pids via child_controller?
   def child_playback_changed(
         _pid,
@@ -142,4 +159,20 @@ defmodule Membrane.Core.Parent.LifecycleController do
 
   defp to_parent_sm_callback(:handle_start_of_stream), do: :handle_element_start_of_stream
   defp to_parent_sm_callback(:handle_end_of_stream), do: :handle_element_end_of_stream
+
+  defp toggle_syncs_active(:prepared, :playing, children_data) do
+    do_toggle_syncs_active(children_data, &Sync.activate/1)
+  end
+
+  defp toggle_syncs_active(:playing, :prepared, children_data) do
+    do_toggle_syncs_active(children_data, &Sync.deactivate/1)
+  end
+
+  defp toggle_syncs_active(_old_playback_state, _new_playback_state, _children_data) do
+    :ok
+  end
+
+  defp do_toggle_syncs_active(children_data, fun) do
+    children_data |> Enum.uniq_by(& &1.sync) |> Enum.map(& &1.sync) |> Bunch.Enum.try_each(fun)
+  end
 end
