+++ lib/membrane/core/pipeline/state.ex
@@ -43,7 +43,7 @@ defmodule Membrane.Core.Pipeline.State do
                 clock_provider: %{clock: nil, provider: nil, choice: :auto},
                 handlers: %{
                   action_handler: Membrane.Pipeline,
-                  playback_controller: Membrane.Pipeline,
+                  playback_controller: Core.Parent.LifecycleController,
                   spec_controller: Core.Pipeline.SpecController
                 }
               ]
